Added draft build process
@stewid

Problems with relative paths to nested /input statements in the
chapters. Last year we used the /input command for each chapter but
reset the /graphicspath before each include in order to find the
figures for each chapter during the full build. This year we have one
level deeper nested /input because of the table and figure.tex files
in each chapter. Currently when we do the build it doesn't work
because of this.

For the fonts, I am using the path of `assets(report)` or
`assets(chapter)` to rewrite the latex with the absolute path to the
fonts on build. I was thinking that maybe a similar approach could be
used. We could rebuild all of the /input statements with the absolute
path when we have that from `report$path` or `chapter$path`. Thoughts
on this?