Functionality to reduce image size
I think that the conditional statements in the arguments to the
to_pdf() function are perhaps flawed. I guess I have to match.arg()
ensure I get just one and not a vector when the argument is
undeclared. Also I'm not sure when the type argument must be specified
and when it is implied in the nexted structure of the to_pdf.chapter,
to_pdf.report and to_pdf set of functions.