Commented out bibliography
Because there are no /citation commands in the Rnw then latex fails
but produces a valid .bbl file as a side effect. Next time you run
latex it uses this .bbl file instead of trying to construct a
bibliography and then how a 0 exit. Either commit the .bbl file or
wait with the bibliography command until we haev citations in the
vignette.