Sample init file
A sample dataset to be used to build a vignette and ensure that we can
build all the way to pdf from the init file. Persently we do not have
a generic method for including the chapters in the pdf build,
therefore the to_pdf of the report will fail on this sample data.