Orgmode added tables and figures
The orgmeode generation is still internal to the package and has no
tool to write it to a file. However, now it searches for the table and
figure files with the functions in the `table.R` and `figure.R` files
and adds subitems to the chapter TODO. I removed the function to do
the formatting and bundled it into the to_opgmode.chapter() function.