Modified parsing of schedule
Instead of identifying the time in the line, look for the keywords
"SCHEDULED" and "DEADLINE". This is easier since there is the
posibility of inserting a time range in the schedule or infact the
deadline. (Although that doesn't make much sense for a deadline). Also
I am using the option invert = NA in the regmatches function whcih was
introduced in R 3.3.0 so I bumped the required verion of R in the
DESCRIPTION file.